pk - google obj exists in layout instead of index
